+++ examples/51Peg/kima_setup.cpp
@@ -10,6 +10,7 @@ const bool obs_after_HARPS_fibers = false;
 const bool GP = false;
 const bool hyperpriors = false;
 const bool trend = false;
+const bool multi_instrument = false;
 
 RVmodel::RVmodel():fix(false),npmax(1)
 {