+++ src/main.cpp
@@ -16,6 +16,7 @@ const bool obs_after_HARPS_fibers = false;
 const bool GP = false;
 const bool hyperpriors = false;
 const bool trend = false;
+const bool multi_instrument = true;
 
 RVmodel::RVmodel():fix(true),npmax(0) {}