+++ tests/test_kima.py
@@ -3,11 +3,13 @@ import os, sys
 import numpy as np
 import numpy.testing as npt
 
+
 def test_import():
     import pykima
 
 
 ## pykima.keplerian
+
 
 def test_true_ecc_anomaly():
     from pykima.keplerian import true_anomaly, ecc_anomaly
@@ -19,23 +21,27 @@ def test_true_ecc_anomaly():
     # eccentric anomaly E
     npt.assert_allclose(ecc_anomaly(0., 0.), 0.)
     npt.assert_allclose(ecc_anomaly(np.pi, 0.), np.pi)
-    npt.assert_allclose(ecc_anomaly(2*np.pi, 0.), 2*np.pi)
+    npt.assert_allclose(ecc_anomaly(2 * np.pi, 0.), 2 * np.pi)
     npt.assert_allclose(ecc_anomaly(0., 0.2), 0.)
     # checked with http://orbitsimulator.com/sheela/kepler.htm
     E = np.rad2deg(ecc_anomaly(np.deg2rad(80), 0.2))
     npt.assert_allclose(E, 91.45545886486815, rtol=1e-8)
 
+
 def test_keplerian():
     from pykima.keplerian import keplerian
     # keplerian(time, p, k, ecc, omega, t0, vsys)
 
-    times1 = [0.,]
+    times1 = [
+        0.,
+    ]
     times2 = [0., 1., 2.]
 
     # before we were testing getting a NaN result when P=0
     # now it should raise a FloatingPointError
     with pytest.warns(RuntimeWarning):
-        with pytest.raises(FloatingPointError, message="Expecting FloatingPointError"):
+        with pytest.raises(FloatingPointError,
+                           message="Expecting FloatingPointError"):
             keplerian(times1, 0., 0., 0., 0., 0., 0.)
             # assert np.isnan(result)
             keplerian(times2, 0., 0., 0., 0., 0., 0.)
@@ -47,6 +53,7 @@ def test_keplerian():
 
 ## pykima.utils
 
+
 def test_percentiles():
     from pykima.utils import percentile68_ranges, percentile68_ranges_latex
 
@@ -57,8 +64,8 @@ def test_percentiles():
     rb = percentile68_ranges(b)
 
     npt.assert_allclose(ra[0], np.median(a))
-    npt.assert_allclose(ra[0]+ra[1], np.percentile(a, 84))
-    npt.assert_allclose(ra[0]-ra[2], np.percentile(a, 16))
+    npt.assert_allclose(ra[0] + ra[1], np.percentile(a, 84))
+    npt.assert_allclose(ra[0] - ra[2], np.percentile(a, 16))
     npt.assert_allclose(ra[0], 0.5, rtol=1e-2)
 
     npt.assert_allclose(rb[0], np.median(b))
@@ -75,17 +82,22 @@ def test_planet_mass():
     from pykima.utils import get_planet_mass
     # get_planet_mass(P, K, e, star_mass=1.0, full_output=False, verbose=False)
     npt.assert_allclose(get_planet_mass(0., 0., 0.), (0., 0.))
-    npt.assert_allclose(get_planet_mass(np.random.rand(), 0., np.random.rand()), 
-                        (0., 0.))
+    npt.assert_allclose(
+        get_planet_mass(np.random.rand(), 0., np.random.rand()), (0., 0.))
+
+
+def test_planet_semimajor_axis():
+    from pykima.utils import get_planet_semimajor_axis
+    # get_planet_semimajor_axis(P, K, star_mass=1.0, full_output=False, verbose=False)
+    pass
 
 
 def test_KimaResults():
     pass
 
 
-
-
 ## pykima.dnest4
+
 
 def test_logsumexp():
     from pykima.classic import logsumexp
@@ -96,6 +108,7 @@ def test_logsumexp():
 
 
 ## pykima.showresults
+
 
 @pytest.fixture(scope='session')
 def simple_results_dir(tmpdir_factory):
@@ -116,17 +129,19 @@ def simple_results_dir(tmpdir_factory):
 
     return directory
 
+
 def write_dummy_model_setup(directory):
     template_setup = ("[kima] \n"
                       "obs_after_HARPS_fibers: false \n"
                       "GP: false \n"
                       "hyperpriors: false \n"
                       "trend: false \n"
-
+                      "multi_instrument: false \n\n"
                       "file: filename.txt \n"
                       "units: kms \n"
                       "skip: 0 \n"
-                     )
+                      "multi: false \n"
+                      "files: \n")
     directory.join('kima_model_setup.txt').write(template_setup)
 
 
@@ -145,20 +160,18 @@ def test_showresults_fails(tmpdir, simple_results_dir, capfd):
     with pytest.raises(IndexError):
         showresults()
 
-    # go to the dummy results directory, which has the 3 files, 
+    # go to the dummy results directory, which has the 3 files,
     # but no kima_model_setup.txt
     os.chdir(str(simple_results_dir))
     if sys.version_info < (3, 0):
-        with pytest.raises(IOError,
-            match="kima_model_setup.txt"):
+        with pytest.raises(IOError, match="kima_model_setup.txt"):
             showresults()
             # out, err = capfd.readouterr()
             write_dummy_model_setup(simple_results_dir)
             with pytest.raises(IOError, match='filename.txt'):
                 showresults()
     else:
-        with pytest.raises(FileNotFoundError,
-            match="kima_model_setup.txt"):
+        with pytest.raises(FileNotFoundError, match="kima_model_setup.txt"):
             showresults()
             # out, err = capfd.readouterr()